Added status codes for each crud request / exception handled

+++ src/main/java/com/nighteye/springbootcrud/controller/UserController.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 import com.nighteye.springbootcrud.model.User;
 import com.nighteye.springbootcrud.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,37 +22,61 @@ public class UserController {
     private UserRepository userRepository;
 
     @GetMapping("/users")
-    private List<User> getAllUsers()
+    private ResponseEntity<List<User>> getAllUsers()
     {
-        List<User> users = new ArrayList<>();
-        userRepository.findAll().forEach(user -> users.add(user));
-        if(users.isEmpty()){
-            return users;
+        try {
+            List<User> users = new ArrayList<>();
+            userRepository.findAll().forEach(user -> users.add(user));
+            if (users.isEmpty()) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
+            return new ResponseEntity<>(users, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return users;
     }
 
     @GetMapping("/users/{userid}")
-    private User getUsers(@PathVariable("userid") int userid)
+    private ResponseEntity<User> getUserById(@PathVariable("userid") int userid)
     {
-        return userRepository.findById(userid).get();
+        Optional<User> user = userRepository.findById(userid);
+        if (user.isPresent()) {
+            return new ResponseEntity<>(user.get(), HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/users/{userid}")
-    private void deleteUser(@PathVariable("userid") int userid)
+    private ResponseEntity<User> deleteUser(@PathVariable("userid") int userid)
     {
-        userRepository.deleteById(userid);
+        try {
+            if(userRepository.findById(userid).isPresent()) {
+                userRepository.deleteById(userid);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PostMapping("/users")
-    private String saveUser(@RequestBody User user)
+    private ResponseEntity<User> saveUser(@RequestBody User user)
     {
-        userRepository.save(user);
-        return user.toString();
+        try {
+            User tempUser = userRepository.save(user);
+            return new ResponseEntity<>(tempUser, HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PutMapping("/users/{userid}")
-    private String update(@PathVariable("userid") int userid, @RequestBody User user)
+    private ResponseEntity<User> update(@PathVariable("userid") int userid, @RequestBody User user)
     {
         Optional<User> userOpt = userRepository.findById(userid);
         if(userOpt.isPresent()) {
@@ -60,9 +86,9 @@ public class UserController {
             tempUser.setEmail(user.getEmail());
             tempUser.setSalary(user.getSalary());
             userRepository.save(tempUser);
-            return tempUser.toString();
+            return new ResponseEntity<>(userRepository.save(tempUser), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        else
-            return "No user with given id found!";
     }
 }
